Implement abstract models.Synthesizer and concrete models.EspeakSynthesizer and models.FestivalSynthesizer
