+++ src/main/java/models/FestivalSynthesizer.java
@@ -1,0 +1,16 @@
+package models;
+
+import java.io.File;
+
+public class FestivalSynthesizer extends Synthesizer {
+
+    @Override
+    public void preview(String text) {
+
+    }
+
+    @Override
+    public void save(String text, File folder) {
+
+    }
+}
